[Security] Upgrade of jackson-databind to resolve CVE-2019-12814
Fixes CVE-2019-12814: A Polymorphic Typing issue was discovered in
FasterXML jackson-databind 2.x through 2.9.9. When Default Typing
is enabled (either globally or for a specific property) for an
externally exposed JSON endpoint and the service has JDOM 1.x or
2.x jar in the classpath, an attacker can send a specifically
crafted JSON message that allows them to read arbitrary local files
on the server.